+++ CHANGELOG.md
@@ -43,6 +43,7 @@ Current
 - [Fix MetricLoader Interface and DimensionalFilter Constructor](https://github.com/yahoo/fili/pull/335)
     * Fix `MetricLoader` interface so that user are not forced to implement deprecated method
     * Fix `DimensionFilter` constructor to not take `null` as a valid dimension and throw a exception instead
+    * Fix `BaseTableLoader` to remove an extra unnecessary `put` to dimension dictionary
 
 - [Fix availability testing utils to be compatible with composite tables](https://github.com/yahoo/fili/pull/419)
     * Fix availability testing utils `populatePhysicalTableCacheIntervals` to assign a `TestAvailability` that will serialize correctly instead of always `StrictAvailability`